DON-1065: Move payment methods out from My Account to a separate page
Not sure why test is currently failing
