+++ src/app/my-account/my-account.component.html
@@ -47,81 +47,6 @@
     </biggive-page-section>
 
     @if (person) {
-      <biggive-page-section>
-        <h2>Your Payment Methods</h2>
-        @if (hasDonationFunds) {
-          <h3>Donation Funds</h3>
-        }
-        @if (hasDonationFunds) {
-          <div>
-            <p>Available balance: <strong>{{ ((person.cash_balance?.gbp || 0) / 100) | exactCurrency:'GBP' }}</strong></p>
-          </div>
-        }
-        @if (registerErrorDescription) {
-          <p class="error" aria-live="assertive">
-            <fa-icon [icon]="faExclamationTriangle"></fa-icon>
-            {{ registerErrorDescription }}
-          </p>
-        }
-        @if (registerSucessMessage) {
-          <p class="update-success" aria-live="assertive">
-            {{ registerSucessMessage }}
-          </p>
-        }
-        @if (paymentMethods === undefined) {
-          <div>
-            <mat-spinner color="accent" [diameter]="40" aria-label="Loading payment methods"></mat-spinner>
-          </div>
-        }
-        <h3>
-          Saved Cards
-        </h3>
-        @if (hasSavedPaymentMethods) {
-          <table id="paymentMethods">
-            @for (method of paymentMethods; track method.id) {
-              <tr>
-                <td class="cardBrand">
-                  <div class="cardBrand">{{(method.card?.brand || method.type).toUpperCase()}}</div>
-                  @if (method.card !== undefined) {
-                    <div>
-                      Card Ending: {{method.card.last4}}
-                    </div>
-                  }
-                  @if (method.card !== undefined) {
-                    <div class="cardExpiry">
-                      Expiry Date: {{method.card.exp_month.toString().padStart(2, "0")}}/{{method.card.exp_year}}
-                    </div>
-                  }
-                </td>
-                <td class="cardExpiry">
-                  @if (method.card !== undefined) {
-                    <div>
-                      Expiry Date: {{method.card.exp_month.toString().padStart(2, "0")}}/{{method.card.exp_year}}
-                    </div>
-                  }
-                </td>
-                @if (method.card !== undefined) {
-                  <td class="cardUpdateButton"
-                    >
-                    <!--            href must be set to make link interactable with keyboard-->
-                    <a href="javascript:void(0);" (click)="updateCard(method.id, method.card, method.billing_details)">
-                      Edit
-                    </a>
-                  </td>
-                }
-                <td class="cardDeleteButton">
-                  <a href="javascript:void(0);" (click)="deleteMethod(method)">Delete</a>
-                </td>
-              </tr>
-            }
-          </table>
-        }
-        @if (paymentMethods !== undefined && paymentMethods.length === 0) {
-          <div>
-            No saved cards
-          </div>
-        }
-      </biggive-page-section>
     }
   </div>
 </main>
