+++ src/app/my-account/my-account.component.spec.ts
@@ -4,12 +4,10 @@ import {
   withInterceptorsFromDi,
 } from "@angular/common/http";
 import { NO_ERRORS_SCHEMA } from "@angular/core";
-import { MatDialogModule } from "@angular/material/dialog";
 import { InMemoryStorageService } from "ngx-webstorage-service";
 import { of } from "rxjs";
-import { PaymentMethod } from "@stripe/stripe-js";
 
-import { DonationService, TBG_DONATE_STORAGE } from "../donation.service";
+import { TBG_DONATE_STORAGE } from "../donation.service";
 import { IdentityService, TBG_DONATE_ID_STORAGE } from "../identity.service";
 import { MyAccountComponent } from "./my-account.component";
 import { ActivatedRoute } from "@angular/router";
@@ -25,7 +23,6 @@ describe("MyAccountComponent", () => {
       declarations: [MyAccountComponent],
       schemas: [NO_ERRORS_SCHEMA],
       imports: [
-        MatDialogModule,
         MatomoModule.forRoot({
           siteId: "",
           trackerUrl: "",
@@ -51,41 +48,6 @@ describe("MyAccountComponent", () => {
       })
     );
 
-    const paymentMethods: PaymentMethod[] = [
-      {
-        created: 0,
-        customer: null,
-        livemode: false,
-        metadata: {},
-        type: "card",
-        object: "payment_method",
-        id: "method_id",
-        billing_details: {
-          address: null,
-          phone: null,
-          email: null,
-          name: null,
-        },
-        card: {
-          networks: null,
-          checks: null,
-          country: null,
-          exp_month: 5,
-          exp_year: 2030,
-          last4: "1234",
-          funding: "",
-          three_d_secure_usage: null,
-          brand: "AcmeCard",
-          wallet: null,
-        },
-      },
-    ];
-
-    const mockDonationSservice = TestBed.inject(DonationService);
-    spyOn(mockDonationSservice, "getPaymentMethods").and.returnValue(
-      of({ data: paymentMethods })
-    );
-
     fixture = TestBed.createComponent(MyAccountComponent);
     component = fixture.componentInstance;
 
@@ -98,11 +60,5 @@ describe("MyAccountComponent", () => {
   it("should show donor name and email", () => {
     expect(element.textContent).toContain("Generous Donor");
     expect(element.textContent).toContain("donor@example.com");
-  });
-
-  it("should show payment card number and brand", () => {
-    expect(element.textContent).toContain(
-      "ACMECARD Card Ending: 1234  Expiry Date: 05/2030"
-    );
   });
 });
