+++ src/app/my-payment-methods/my-payment-methods.component.cy.ts
@@ -1,0 +1,7 @@
+import { MyPaymentMethodsComponent } from './my-payment-methods.component'
+
+describe('MyPaymentMethodsComponent', () => {
+  it('should mount', () => {
+    cy.mount(MyPaymentMethodsComponent)
+  })
+})
