+++ src/app/my-payment-methods/my-payment-methods.component.html
@@ -1,0 +1,85 @@
+<div id="banner"></div>
+<main>
+  <div>
+    <biggive-page-section>
+      <div class="heading-wrapper">
+        <biggive-heading
+          colour="tertiary"
+          size="1"
+          align="center"
+          text="Your Payment Methods"
+        ></biggive-heading>
+      </div>
+        <h2>Your Payment Methods</h2>
+          <h3>Donation Funds</h3>
+          <div>
+            @if (hasDonationFunds) {
+            <p>Available balance: <strong>{{ ((person.cash_balance?.gbp || 0) / 100) | exactCurrency:'GBP' }}</strong></p>
+            } @else {
+              <p>No donation funds</p>
+            }
+            <biggive-button
+              space-above="5"
+              colour-scheme="white"
+              label="Transfer funds"
+              full-width="false"
+              size="medium"
+              url="/transfer-funds"
+            />
+          </div>
+        <h3>
+          Saved Cards
+        </h3>
+      @if (paymentMethods === undefined) {
+        <div>
+          <mat-spinner color="accent" [diameter]="40" aria-label="Loading payment methods"></mat-spinner>
+        </div>
+      }
+      @if (hasSavedPaymentMethods) {
+          <table id="paymentMethods">
+            @for (method of paymentMethods; track method.id) {
+              <tr>
+                <td class="cardBrand">
+                  <div class="cardBrand">{{(method.card?.brand || method.type).toUpperCase()}}</div>
+                  @if (method.card !== undefined) {
+                    <div>
+                      Card Ending: {{method.card.last4}}
+                    </div>
+                  }
+                  @if (method.card !== undefined) {
+                    <div class="cardExpiry">
+                      Expiry Date: {{method.card.exp_month.toString().padStart(2, "0")}}/{{method.card.exp_year}}
+                    </div>
+                  }
+                </td>
+                <td class="cardExpiry">
+                  @if (method.card !== undefined) {
+                    <div>
+                      Expiry Date: {{method.card.exp_month.toString().padStart(2, "0")}}/{{method.card.exp_year}}
+                    </div>
+                  }
+                </td>
+                @if (method.card !== undefined) {
+                  <td class="cardUpdateButton"
+                  >
+                    <!--            href must be set to make link interactable with keyboard-->
+                    <a href="javascript:void(0);" (click)="updateCard(method.id, method.card, method.billing_details)">
+                      Edit
+                    </a>
+                  </td>
+                }
+                <td class="cardDeleteButton">
+                  <a href="javascript:void(0);" (click)="deleteMethod(method)">Delete</a>
+                </td>
+              </tr>
+            }
+          </table>
+        }
+        @if (paymentMethods !== undefined && paymentMethods.length === 0) {
+          <div>
+            No saved cards
+          </div>
+        }
+    </biggive-page-section>
+  </div>
+</main>
